Move token_module into protocol_level_tokens
