+++ plt/plt-scheduler/src/protocol_level_tokens.rs
@@ -1,3 +1,4 @@
 pub mod balance_operations;
 pub mod p11;
 pub mod p9;
+pub mod token_module;
