--- plt/plt-scheduler/src/token_module/module/mod.rs
@@ -1,3 +1,0 @@
-pub mod initialize;
-pub mod queries;
-pub mod update;
